+++ test/CommandLineUtils.Tests/Utilities/TestConsole.cs
@@ -3,6 +3,8 @@
 
 using System;
 using System.IO;
+using System.Linq;
+using System.Reflection;
 using Xunit.Abstractions;
 
 namespace McMaster.Extensions.CommandLineUtils.Tests
@@ -38,7 +40,12 @@ namespace McMaster.Extensions.CommandLineUtils.Tests
 
         public void RaiseCancelKeyPress()
         {
-            CancelKeyPress?.Invoke(this, default);
+            // See https://github.com/dotnet/corefx/blob/f2292af3a1794378339d6f5c8adcc0f2019a2cf9/src/System.Console/src/System/ConsoleCancelEventArgs.cs#L14
+            var eventArgs = typeof(ConsoleCancelEventArgs)
+                .GetConstructors(BindingFlags.NonPublic | BindingFlags.Instance)
+                .First()
+                .Invoke(new object[] { ConsoleSpecialKey.ControlC });
+            CancelKeyPress?.Invoke(this, (ConsoleCancelEventArgs)eventArgs);
         }
     }
 }